rgw/d3n: move d3n_cache_op to rgw_d3n_cacherequest.*
moves the d3n cache op from rgw_aio.* to rgw_d3n_cacherequest.*

removes empty class D3nL1CacheRequest which we allocated on the heap
unnecessarily. its member functions are now free functions
