+++ src/rgw/CMakeLists.txt
@@ -74,6 +74,7 @@ set(librgw_common_srcs
   rgw_compression.cc
   rgw_cors.cc
   rgw_cors_s3.cc
+  rgw_d3n_cacherequest.cc
   rgw_env.cc
   rgw_es_query.cc
   rgw_formats.cc
