+++ src/rgw/rgw_aio.cc
@@ -18,7 +18,6 @@
 #include "librados/librados_asio.h"
 
 #include "rgw_aio.h"
-#include "rgw_d3n_cacherequest.h"
 
 namespace rgw {
 
@@ -101,18 +100,6 @@ Aio::OpFunc aio_abstract(librados::IoCtx ctx, Op&& op,
     };
 }
 
-
-Aio::OpFunc d3n_cache_aio_abstract(const DoutPrefixProvider *dpp, optional_yield y, off_t read_ofs, off_t read_len, std::string& cache_location) {
-  return [dpp, y, read_ofs, read_len, cache_location] (Aio* aio, AioResult& r) mutable {
-    // d3n data cache requires yield context (rgw_beast_enable_async=true)
-    ceph_assert(y);
-    auto c = std::make_unique<D3nL1CacheRequest>();
-    lsubdout(g_ceph_context, rgw_datacache, 20) << "D3nDataCache: d3n_cache_aio_abstract(): libaio Read From Cache, oid=" << r.obj.oid << dendl;
-    c->file_aio_read_abstract(dpp, y.get_io_context(), y.get_yield_context(), cache_location, read_ofs, read_len, aio, r);
-  };
-}
-
-
 template <typename Op>
 Aio::OpFunc aio_abstract(librados::IoCtx ctx, Op&& op, optional_yield y) {
   static_assert(std::is_base_of_v<librados::ObjectOperation, std::decay_t<Op>>);
@@ -136,11 +123,6 @@ Aio::OpFunc Aio::librados_op(librados::IoCtx ctx,
                              librados::ObjectWriteOperation&& op,
                              optional_yield y) {
   return aio_abstract(std::move(ctx), std::move(op), y);
-}
-
-Aio::OpFunc Aio::d3n_cache_op(const DoutPrefixProvider *dpp, optional_yield y,
-                              off_t read_ofs, off_t read_len, std::string& cache_location) {
-  return d3n_cache_aio_abstract(dpp, y, read_ofs, read_len, cache_location);
 }
 
 } // namespace rgw
