+++ src/rgw/rgw_aio.h
@@ -97,8 +97,6 @@ class Aio {
   static OpFunc librados_op(librados::IoCtx ctx,
                             librados::ObjectWriteOperation&& op,
                             optional_yield y);
-  static OpFunc d3n_cache_op(const DoutPrefixProvider *dpp, optional_yield y,
-                             off_t read_ofs, off_t read_len, std::string& location);
 };
 
 } // namespace rgw
